+++ Data/SportBattles.Data.Models/Prediction.cs
@@ -20,9 +20,9 @@
 
         public Match Match { get; set; }
 
-        public int? HomeGoals { get; set; }
+        public byte? HomeGoals { get; set; }
 
-        public int? AwayGoals { get; set; }
+        public byte? AwayGoals { get; set; }
 
         public ThreeWayResult? ThreeWayWinner { get; set; }
     }
